requirements.txt
Django==5.2.5
djangorestframework
djangorestframework-simplejwt
python-decouple
psycopg2-binary
drf-spectacular